move tests relative to from_marbles()/cold() and hot() to tests/test_observable/test_marbles.py
